Changed wavenumber data type to float64
I'm not sure whether this has expected (for you) or unexpected (for me) side effects, but this is how I fixed it locally. I don't see how to distinguish the data type based on what the dtype of the ParticleMesh is.. please consider this PR more as a hotfix instead of a carefully designed feature extension.